% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inteiroTJAL.R
\name{inteiroTJAL}
\alias{inteiroTJAL}
\title{This function downloads complete decisions from Alagoas state High Court.}
\usage{
inteiroTJAL(url = NULL, cdacordao = NULL)
}
\arguments{
\item{url}{a vector of url strings to access the decisions.}

\item{cdacordao}{a vector of decisions ids/codes to give as names to the downloading files.}
}
\value{
pdf of the decision downloaded to the working directory.
}
\description{
This function downloads complete decisions from Alagoas state High Court.
}
\keyword{court}
\keyword{decisions}
\keyword{tjal}
