% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inteiroTJAL.R
\name{testar_processoAL}
\alias{testar_processoAL}
\title{Função inteiroTJAL}
\usage{
testar_processoAL(url)
}
\arguments{
\item{url}{a vector of url strings to access the decisions.}
}
\value{
pdf of the decision downloaded to the working directory.
}
\description{
This function predicts captcha letters
}
\keyword{court}
\keyword{decisions}
\keyword{tjsp}
