% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjalsg_meta.R
\name{tjalsg_meta}
\alias{tjalsg_meta}
\title{Função tjalsg_meta}
\usage{
tjalsg_meta(BuscaLivre, quote = TRUE, classes.value = "", inicio = "",
  fim = "")
}
\arguments{
\item{quote}{logical. Colocar a expressão entre aspas?}

\item{classes.value}{Código}

\item{inicio}{Data inicial}

\item{fim}{Data final}

\item{livre}{palavra ou texto a ser buscado nas ementas e nos acórdãos}
}
\description{
Esta função extrai metadados das decisões de segundo grau do TJAL
}
\examples{
tjalsg_meta(BuscaLivre="Lei Maria da Penha")
}
\keyword{alagoas}
\keyword{court,}
\keyword{decisions,}
\keyword{high}
\keyword{judicial}
\keyword{tjal,}
